+++ docs/changes.rst
@@ -11,6 +11,7 @@ Released on ? 2017.
 * Added support for blocking certain emails from registration.
 * Users can now delete their own comments.
 * Added preview step to search and replace feature.
+* Client side persistence of settings in search and upload forms.
 
 weblate 2.16
 ------------